web/requirements.txt
asgiref==3.5.2
django==3.2.11
pytz==2022.6
sqlparse==0.4.3
psycopg2==2.9.3
gunicorn==20.1.0

dj-rest-auth==2.2.4
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
drf-yasg==1.20.0 # Swagger
django-cors-headers==3.11.0
django-phonenumber-field==5.2.0
phonenumbers==8.12.21
django-filter==21.1
pillow==8.2.0
channels==4.0.0
channels_redis==4.0.0
redis==4.3.1
uvicorn==0.22.0
websockets>=10.4
httptools>=0.5.0
urllib3==1.26.14
